Refactor: Make HTTP Prober a Subclass (#1107)
* refactor: create implementation of http prober based on base prober

* refactor: make base prober lean by removing unneded code

* refactor: add http prober to factory

* refactor: replace function with subclass

* refactor: inline probe non http function
