+++ src/components/probe/prober/factory.ts
@@ -22,6 +22,7 @@
  * SOFTWARE.                                                                      *
  **********************************************************************************/
 
+import { HTTPProber } from './http'
 import { MongoProber } from './mongo'
 import { MariaDBProber } from './mariadb'
 import { PostgresProber } from './postgres'
@@ -32,6 +33,10 @@ import { BaseProber, type Prober, type ProberMetadata } from '.'
 export function createProbers(probeMetadata: ProberMetadata): Prober[] {
   const { probeConfig } = probeMetadata
   const result: Prober[] = []
+
+  if (probeConfig?.requests) {
+    result.push(createProber(probeMetadata))
+  }
 
   if (probeConfig?.mongo) {
     result.push(createProber(probeMetadata))
@@ -58,6 +63,10 @@ export function createProbers(probeMetadata: ProberMetadata): Prober[] {
 
 export function createProber(probeMetadata: ProberMetadata): Prober {
   const { probeConfig } = probeMetadata
+
+  if (probeConfig?.requests) {
+    return new HTTPProber(probeMetadata)
+  }
 
   if (probeConfig?.mariadb || probeConfig?.mysql) {
     return new MariaDBProber(probeMetadata)
